tests/baseline/protocol-1.2/inspector-protocol/css/css-get-background-colors-expected.txt

Lorem ipsum dolor sit amet, consectetur adipiscing elit, sed do eiusmod tempor incididunt ut labore et dolore magna aliqua. Ut enim ad minim veniam, quis nostrud exercitation ullamco laboris nisi ut aliquip ex ea commodo consequat.
Lorem ipsum dolor sit amet, consectetur adipiscing elit, sed do eiusmod tempor incididunt ut labore et dolore magna aliqua. Ut enim ad minim veniam, quis nostrud exercitation ullamco laboris nisi ut aliquip ex ea commodo consequat.
Lorem ipsum dolor sit amet, consectetur adipiscing elit, sed do eiusmod tempor incididunt ut labore et dolore magna aliqua. Ut enim ad minim veniam, quis nostrud exercitation ullamco laboris nisi ut aliquip ex ea commodo consequat.
Lorem ipsum dolor sit amet, consectetur adipiscing elit, sed do eiusmod tempor incididunt ut labore et dolore magna aliqua. Ut enim ad minim veniam, quis nostrud exercitation ullamco laboris nisi ut aliquip ex ea commodo consequat.
Lorem ipsum dolor sit amet, consectetur adipiscing elit, sed do eiusmod tempor incididunt ut labore et dolore magna aliqua. Ut enim ad minim veniam, quis nostrud exercitation ullamco laboris nisi ut aliquip ex ea commodo consequat.
Lorem ipsum dolor sit amet, consectetur adipiscing elit, sed do eiusmod tempor incididunt ut labore et dolore magna aliqua. Ut enim ad minim veniam, quis nostrud exercitation ullamco laboris nisi ut aliquip ex ea commodo consequat.
Lorem ipsum dolor sit amet, consectetur adipiscing elit, sed do eiusmod tempor incididunt ut labore et dolore magna aliqua. Ut enim ad minim veniam, quis nostrud exercitation ullamco laboris nisi ut aliquip ex ea commodo consequat.
Lorem ipsum dolor sit amet, consectetur adipiscing elit, sed do eiusmod tempor incididunt ut labore et dolore magna aliqua. Ut enim ad minim veniam, quis nostrud exercitation ullamco laboris nisi ut aliquip ex ea commodo consequat.
Lorem ipsum dolor sit amet, consectetur adipiscing elit, sed do eiusmod tempor incididunt ut labore et dolore magna aliqua. Ut enim ad minim veniam, quis nostrud exercitation ullamco laboris nisi ut aliquip ex ea commodo consequat.
Lorem ipsum dolor sit amet, consectetur adipiscing elit, sed do eiusmod tempor incididunt ut labore et dolore magna aliqua. Ut enim ad minim veniam, quis nostrud exercitation ullamco laboris nisi ut aliquip ex ea commodo consequat.
Lorem ipsum dolor sit amet, consectetur adipiscing elit, sed do eiusmod tempor incididunt ut labore et dolore magna aliqua. Ut enim ad minim veniam, quis nostrud exercitation ullamco laboris nisi ut aliquip ex ea commodo consequat.
Lorem ipsum dolor sit amet, consectetur adipiscing elit, sed do eiusmod tempor incididunt ut labore et dolore magna aliqua. Ut enim ad minim veniam, quis nostrud exercitation ullamco laboris nisi ut aliquip ex ea commodo consequat.
Lorem ipsum dolor sit amet, consectetur adipiscing elit, sed do eiusmod tempor incididunt ut labore et dolore magna aliqua. Ut enim ad minim veniam, quis nostrud exercitation ullamco laboris nisi ut aliquip ex ea commodo consequat.
Lorem ipsum dolor sit amet, consectetur adipiscing elit, sed do eiusmod tempor incididunt ut labore et dolore magna aliqua. Ut enim ad minim veniam, quis nostrud exercitation ullamco laboris nisi ut aliquip ex ea commodo consequat.
Lorem ipsum dolor sit amet, consectetur adipiscing elit, sed do eiusmod tempor incididunt ut labore et dolore magna aliqua. Ut enim ad minim veniam, quis nostrud exercitation ullamco laboris nisi ut aliquip ex ea commodo consequat.
Lorem ipsum dolor sit amet, consectetur adipiscing elit, sed do eiusmod tempor incididunt ut labore et dolore magna aliqua. Ut enim ad minim veniam, quis nostrud exercitation ullamco laboris nisi ut aliquip ex ea commodo consequat.
Lorem ipsum dolor sit amet, consectetur adipiscing elit, sed do eiusmod tempor incididunt ut labore et dolore magna aliqua. Ut enim ad minim veniam, quis nostrud exercitation ullamco laboris nisi ut aliquip ex ea commodo consequat.
Lorem ipsum dolor sit amet, consectetur adipiscing elit, sed do eiusmod tempor incididunt ut labore et dolore magna aliqua. Ut enim ad minim veniam, quis nostrud exercitation ullamco laboris nisi ut aliquip ex ea commodo consequat.
Lorem ipsum dolor sit amet, consectetur adipiscing elit, sed do eiusmod tempor incididunt ut labore et dolore magna aliqua. Ut enim ad minim veniam, quis nostrud exercitation ullamco laboris nisi ut aliquip ex ea commodo consequat.
Lorem ipsum dolor sit amet, consectetur adipiscing elit, sed do eiusmod tempor incididunt ut labore et dolore magna aliqua. Ut enim ad minim veniam, quis nostrud exercitation ullamco laboris nisi ut aliquip ex ea commodo consequat.
Lorem ipsum dolor sit amet, consectetur adipiscing elit, sed do eiusmod tempor incididunt ut labore et dolore magna aliqua. Ut enim ad minim veniam, quis nostrud exercitation ullamco laboris nisi ut aliquip ex ea commodo consequat.
Lorem ipsum dolor sit amet, consectetur adipiscing elit, sed do eiusmod tempor incididunt ut labore et dolore magna aliqua. Ut enim ad minim veniam, quis nostrud exercitation ullamco laboris nisi ut aliquip ex ea commodo consequat.
Lorem ipsum dolor sit amet, consectetur adipiscing elit, sed do eiusmod tempor incididunt ut labore et dolore magna aliqua. Ut enim ad minim veniam, quis nostrud exercitation ullamco laboris nisi ut aliquip ex ea commodo consequat.
Lorem ipsum dolor sit amet, consectetur adipiscing elit, sed do eiusmod tempor incididunt ut labore et dolore magna aliqua. Ut enim ad minim veniam, quis nostrud exercitation ullamco laboris nisi ut aliquip ex ea commodo consequat.
Lorem ipsum dolor sit amet, consectetur adipiscing elit, sed do eiusmod tempor incididunt ut labore et dolore magna aliqua. Ut enim ad minim veniam, quis nostrud exercitation ullamco laboris nisi ut aliquip ex ea commodo consequat.

Running test: testNoText
No text: should be null
{}

Running test: testNoBgColor
No background color: should be white
{"backgroundColors":["rgb(255, 255, 255)"]}

Running test: testOpaqueBgColor
Opaque background color: should be red
{"backgroundColors":["rgb(255, 0, 0)"]}

Running test: testLayeredOpaqueBgColors
Opaque background color in front of another opaque background color: should be blue
{"backgroundColors":["rgb(0, 0, 255)"]}

Running test: testOneSemitransparentBgColor
Semi-transparent background color: should be light red
{"backgroundColors":["rgb(255, 128, 128)"]}

Running test: testTwoSemitransparentBgColors
Two layered semi-transparent background colors: should be medium red
{"backgroundColors":["rgb(255, 64, 64)"]}

Running test: testOpaqueGradientBackground
Opaque gradient: should be red and black
{"backgroundColors":["rgb(255, 0, 0)","rgb(0, 0, 0)"]}

Running test: testOpaqueGradientBackgroundBehindScrim
Opaque gradient behind semi-transparent color: should be light red and 50% grey
{"backgroundColors":["rgb(255, 127, 127)","rgb(127, 127, 127)"]}

Running test: testOpaqueGradientBackgroundWithColorBackground
Opaque gradient and solid color background on same element: should be red and black
{"backgroundColors":["rgb(255, 0, 0)","rgb(0, 0, 0)"]}

Running test: testPartiallyTransparentGradientBackground
Semi-transparent gradient: should be light red and 50% grey
{"backgroundColors":["rgb(255, 128, 128)","rgb(128, 128, 128)"]}

Running test: testPartiallyTransparentGradientAndColorBackground
Semi-transparent gradient and solid color on same element: should be dark red and 50% grey
{"backgroundColors":["rgb(127, 0, 0)","rgb(127, 127, 127)"]}

Running test: testTwoPartiallyTransparentGradientBackgrounds
Layered semi-transparent gradients: should be empty array
{"backgroundColors":[]}

Running test: testPartiallyOverlappingBackground
Partially overlapping background: should be empty array
{"backgroundColors":[]}

Running test: smallerBackground
Background smaller than text: should be empty array
{"backgroundColors":[]}

Running test: testObscuredPartiallyOverlappingBackground
Red background obscuring partially overlapping blue background: should be red only
{"backgroundColors":["rgb(255, 0, 0)"]}

Running test: testBackgroundImage
Background image: should be empty array
{"backgroundColors":[]}

Running test: testBackgroundImageAndBgColor
Background image with background color: should be empty array
{"backgroundColors":[]}

Running test: testBackgroundImageBehindScrim
Background image behind scrim: should be semi-transparent white
{"backgroundColors":["rgba(255, 255, 255, 0.498039)"]}

Running test: testForegroundImage
Image behind text: should be empty array
{"backgroundColors":[]}

Running test: testForegroundImageBehindScrim
Image behind scrim: should be semi-transparent white
{"backgroundColors":["rgba(255, 255, 255, 0.498039)"]}

Running test: testForegroundCanvas
Canvas behind text: should be empty array
{"backgroundColors":[]}

Running test: testForegroundEmbed
Embed behind text: should be empty array
{"backgroundColors":[]}

Running test: testForegroundObject
Object behind text: should be empty array
{"backgroundColors":[]}

Running test: testForegroundPicture
Picture behind text: should be empty array
{"backgroundColors":[]}

Running test: testForegroundSVG
SVG behind text: should be empty array
{"backgroundColors":[]}

Running test: testForegroundVideo
Video behind text: should be empty array
{"backgroundColors":[]}
